+++ src/main/java/domain/Lines.java
@@ -2,7 +2,7 @@ package domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import domain.booleangenerator.BooleanGenerator;
+import domain.generator.BooleanGenerator;
 
 public class Lines {
     private final List<Line> lines = new ArrayList<>();
