+++ src/test/java/domain/LineTest.java
@@ -5,9 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import domain.booleangenerator.BooleanGenerator;
-import domain.booleangenerator.RandomBooleanGenerator;
-import domain.booleangenerator.TrueBooleanGenerator;
+import domain.generator.BooleanGenerator;
+import domain.generator.RandomBooleanGenerator;
+import domain.generator.TrueBooleanGenerator;
 
 class LineTest {
     private final BooleanGenerator randomBooleanGenerator = new RandomBooleanGenerator();
